fixing Article.java

+++ app/src/main/java/com/example/proxymeister/antonsskafferi/model/Article.java
@@ -1,20 +1,16 @@
 package com.example.proxymeister.antonsskafferi.model;
 
-/**
- * Created by curious on 2015-10-08.
- */
 public class Article {
     public int id;
     public String name;
     // public String category; unlikely used
-    public int amount;
+    public double amount;
     public String unit;
     public String exp_date;     // public String days_to_exp ( depending on solution )
 
     @Override
     public String toString()
     {
-        return name + " | " + Integer.toString(amount) + " " + unit + " | " +  exp_date;
+        return name + " | " + Double.toString(amount) + " " + unit + " | " +  exp_date;
     }
 }
-
